+++ backend/clickhouse/sessions.go
@@ -108,6 +108,7 @@ type ClickhouseField struct {
 	SessionCreatedAt int64
 	SessionID        int64
 	Value            string
+	Timestamp        int64
 }
 
 // These keys show up as recommendations, not in fields table due to high cardinality or post processing booleans
@@ -125,6 +126,7 @@ var defaultSessionsKeys = []*modelInputs.QueryKey{
 	{Name: string(modelInputs.ReservedSessionKeySecureID), Type: modelInputs.KeyTypeString},
 	{Name: string(modelInputs.ReservedSessionKeyViewedByAnyone), Type: modelInputs.KeyTypeBoolean},
 	{Name: string(modelInputs.ReservedSessionKeyViewedByMe), Type: modelInputs.KeyTypeBoolean},
+	{Name: string(modelInputs.ReservedSessionKeyTimestamp), Type: modelInputs.KeyTypeNumeric},
 }
 
 var booleanKeys = map[string]bool{
@@ -177,6 +179,7 @@ func (client *Client) WriteSessions(ctx context.Context, sessions []*model.Sessi
 				Value:            field.Value,
 				SessionID:        int64(session.ID),
 				SessionCreatedAt: session.CreatedAt.UnixMicro(),
+				Timestamp:        field.Timestamp.UnixMicro(),
 			}
 			chFields = append(chFields, &chf)
 		}
@@ -521,8 +524,10 @@ var SessionsJoinedTableConfig = model.TableConfig{
 		string(modelInputs.ReservedSessionKeyPagesVisited):       "PagesVisited",
 		string(modelInputs.ReservedSessionKeySecureID):           "SecureID",
 		string(modelInputs.ReservedSessionKeyState):              "State",
+		string(modelInputs.ReservedSessionKeyTimestamp):          "Timestamp",
 		string(modelInputs.ReservedSessionKeyViewedByAnyone):     "Viewed",
 		string(modelInputs.ReservedSessionKeyWithinBillingQuota): "WithinBillingQuota",
+		string(modelInputs.ReservedSessionKeyUpdatedAt):          "UpdatedAt",
 
 		// deprecated but kept in for backwards compatibility of search
 		string(modelInputs.ReservedSessionKeyViewed):    "Viewed",