+++ highlight.io/components/QuickstartContent/traces/node-js/nextjs.tsx
@@ -26,7 +26,7 @@ export default withPageRouterHighlight(async function handler(
 	res: NextApiResponse,
 ) {
 	return new Promise<void>(async (resolve) => {
-		const span = await H.startActiveSpan('page-router-span', {})
+		const { span } = H.startWithHeaders('page-router-span', {})
 
 		console.info('Here: /pages/api/page-router-trace.ts ⌚⌚⌚')
 
@@ -54,7 +54,7 @@ export const GET = withAppRouterHighlight(async function GET(
 	request: NextRequest,
 ) {
 	return new Promise(async (resolve) => {
-		const span = await H.startActiveSpan('app-router-span', {})
+		const { span } = H.startWithHeaders('app-router-span', {})
 
 		console.info('Here: /pages/api/app-router-trace/route.ts ⏰⏰⏰')